+++ examples/instancing.rs
@@ -320,7 +320,7 @@ impl RenderState {
             total_instances: NUM_QUADS,
         };
 
-        self.params_deposit.write_data(0, &[params])?;
+        (&self.params_deposit << &params)?;
         self.upload_scheme.submit()?;
 
         let mut submission = self.scheme.submit()?;