+++ examples/solid_cube.rs
@@ -366,14 +366,8 @@ impl App {
         }
 
         let upload = self.upload_scheme.as_mut().unwrap();
-        self.vertex_deposit
-            .as_ref()
-            .unwrap()
-            .write(0, bytemuck::cast_slice(&vertices))?;
-        self.index_deposit
-            .as_ref()
-            .unwrap()
-            .write(0, bytemuck::cast_slice(&sorted_indices))?;
+        (self.vertex_deposit.as_ref().unwrap() << vertices.as_slice())?;
+        (self.index_deposit.as_ref().unwrap() << sorted_indices.as_slice())?;
         upload.submit()?;
 
         let scheme = self.scheme.as_mut().unwrap();